WW-4762 Moves common method up
